vibetrackr/backend/requirements.txt
accelerate==1.7.0
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.2
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
dnspython==2.7.0
filelock==3.18.0
firebase-admin==6.8.0
flasgger==0.9.7.1
Flask==3.1.1
flask-cors==6.0.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-genai==1.16.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio_client==1.10.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.6
packaging==23.2
pandas==2.2.3
pillow==11.2.1
protobuf==5.29.4
psutil==5.9.5
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.13.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
spotapi==1.1.8
spotipy==2.25.1
sympy==1.14.0
threadpoolctl==3.6.0
tls-client==1.0.1
tokenizers==0.13.3
torch==2.2.2
tqdm==4.67.1
transformers==4.29.0
typing-inspection==0.4.1
typing_extensions==4.11.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
validators==0.35.0
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.0
gunicorn